+++ DecoratorFactory/src/calculator/CalculableFactory.java
@@ -1,7 +1,7 @@
 package calculator;
 
 public class CalculableFactory implements ICalculableFactory {
-    public Calculable create(int primaryArg) {
+    public Calculable create(float primaryArg) {
         return new Calculator(primaryArg);
     }
 }
